yaida: renamed RepoManager.cpp/hpp to Repository.cpp/hpp
