+++ YaLibs/YaToolsIDALib/Repository.cpp
@@ -13,7 +13,7 @@
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "RepoManager.hpp"
+#include "Repository.hpp"
 
 #include "../YaGitLib/YaGitLib.hpp"
 #include "../YaGitLib/ResolveFileConflictCallback.hpp"
