+++ YaLibs/YaToolsPy/YaToolsPy.h
@@ -54,7 +54,7 @@ namespace std
 #include "Model.hpp"
 #include "MultiplexerDelegatingVisitor.hpp"
 #include "PathDebuggerVisitor.hpp"
-#include "RepoManager.hpp"
+#include "Repository.hpp"
 #include "ResolveFileConflictCallback.hpp"
 #include "Signature.hpp"
 #include "XML/XMLDatabaseModel.hpp"
@@ -201,7 +201,7 @@ namespace std
 %include "Model.hpp"
 %include "MultiplexerDelegatingVisitor.hpp"
 %include "PathDebuggerVisitor.hpp"
-%include "RepoManager.hpp"
+%include "Repository.hpp"
 %include "ResolveFileConflictCallback.hpp"
 %include "XML/XMLDatabaseModel.hpp"
 %include "XML/XMLExporter.hpp"
